+++ beacon-chain/state/interface/beacon_state_v0.go
